refactor: update tests after moving routes
